Move response path of AI APIs to AIAPIMediator
