+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/mediators/AIAPIMediator.java
@@ -19,27 +19,37 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.synapse.ManagedLifecycle;
 import org.apache.synapse.MessageContext;
+import org.apache.synapse.commons.json.JsonUtil;
 import org.apache.synapse.core.SynapseEnvironment;
 import org.apache.synapse.core.axis2.Axis2MessageContext;
 import org.apache.synapse.mediators.AbstractMediator;
+import org.apache.synapse.rest.RESTConstants;
+import org.apache.synapse.transport.passthru.util.RelayUtils;
 import org.wso2.carbon.apimgt.api.APIConstants;
+import org.wso2.carbon.apimgt.api.APIManagementException;
 import org.wso2.carbon.apimgt.api.LLMProviderConfiguration;
+import org.wso2.carbon.apimgt.api.LLMProviderMetadata;
 import org.wso2.carbon.apimgt.api.LLMProviderService;
 import org.wso2.carbon.apimgt.api.model.LLMProviderInfo;
+import org.wso2.carbon.apimgt.gateway.APIMgtGatewayConstants;
 import org.wso2.carbon.apimgt.gateway.internal.DataHolder;
 import org.wso2.carbon.apimgt.gateway.internal.ServiceReferenceHolder;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
+import javax.ws.rs.core.MediaType;
+import javax.xml.stream.XMLStreamException;
 
 /**
- * AIAPIMediator is responsible for extracting payload metadata from AI API requests
+ * AIAPIMediator is responsible for extracting payload metadata from AI API requests/response
  * and setting it in the message context for further processing.
  */
 public class AIAPIMediator extends AbstractMediator implements ManagedLifecycle {
 
     private static final Log log = LogFactory.getLog(AIAPIMediator.class);
     private String llmProviderId;
+    private String direction;
 
     /**
      * Initializes the AIAPIMediator.
@@ -75,28 +85,45 @@ public class AIAPIMediator extends AbstractMediator implements ManagedLifecycle 
         if (log.isDebugEnabled()) {
             log.debug("AIAPIMediator: Mediation started.");
         }
+        try {
+            LLMProviderInfo provider = DataHolder.getInstance().getLLMProviderConfigurations(this.llmProviderId);
+            if (provider == null) {
+                log.error("No LLM provider found for provider ID: " + llmProviderId);
+                return false;
+            }
 
-        LLMProviderInfo provider = DataHolder.getInstance().getLLMProviderConfigurations(this.llmProviderId);
-        if (provider == null) {
-            log.error("No LLM provider found for provider ID: " + llmProviderId);
+            LLMProviderConfiguration providerConfiguration = provider.getConfigurations();
+            LLMProviderService llmProviderService = ServiceReferenceHolder.getInstance()
+                    .getLLMProviderService(providerConfiguration.getConnectorType());
+
+            if (llmProviderService == null) {
+                log.error("LLM provider service not found for the provider with ID: " + llmProviderId);
+                return false;
+            }
+
+            Map<String, String> metadataMap = new HashMap<>();
+            if (APIConstants.AIAPIConstants.TRAFFIC_FLOW_DIRECTION_IN.equals(direction)) {
+                metadataMap.put(APIConstants.AIAPIConstants.NAME, provider.getName());
+                metadataMap.put(APIConstants.AIAPIConstants.API_VERSION, provider.getApiVersion());
+                ((Axis2MessageContext) messageContext).getAxis2MessageContext()
+                        .setProperty(APIConstants.AIAPIConstants.AI_API_REQUEST_METADATA, metadataMap);
+            } else if (APIConstants.AIAPIConstants.TRAFFIC_FLOW_DIRECTION_OUT.equals(direction)) {
+                String payload = extractPayloadFromContext(messageContext, providerConfiguration);
+                Map<String, String> queryParams = extractQueryParamsFromContext(messageContext);
+                Map<String, String> headers = extractHeadersFromContext(messageContext);
+                llmProviderService.getResponseMetadata(payload, headers, queryParams,
+                        providerConfiguration.getMetadata(), metadataMap);
+                ((Axis2MessageContext) messageContext).getAxis2MessageContext()
+                        .setProperty(APIConstants.AIAPIConstants.AI_API_RESPONSE_METADATA, metadataMap);
+            }
+        } catch (Exception e) {
+            if (e instanceof APIManagementException) {
+                log.error("Error occurred while extracting metadata.", e);
+            } else if (e instanceof XMLStreamException || e instanceof IOException) {
+                log.error("Error occurred while reading payload.", e);
+            }
             return false;
         }
-
-        LLMProviderConfiguration providerConfiguration = provider.getConfigurations();
-        LLMProviderService llmProviderService = ServiceReferenceHolder.getInstance()
-                .getLLMProviderService(providerConfiguration.getConnectorType());
-
-        if (llmProviderService == null) {
-            log.error("LLM provider service not found for the provider with ID: " + llmProviderId);
-            return false;
-        }
-
-        Map<String, String> metadataMap = new HashMap<>();
-        metadataMap.put(APIConstants.AIAPIConstants.NAME, provider.getName());
-        metadataMap.put(APIConstants.AIAPIConstants.API_VERSION, provider.getApiVersion());
-
-        ((Axis2MessageContext) messageContext).getAxis2MessageContext()
-                .setProperty(APIConstants.AIAPIConstants.AI_API_REQUEST_METADATA, metadataMap);
 
         return true;
     }
@@ -130,5 +157,134 @@ public class AIAPIMediator extends AbstractMediator implements ManagedLifecycle 
     public void setLlmProviderId(String llmProviderId) {
 
         this.llmProviderId = llmProviderId;
+    }
+
+    /**
+     * Retrieves the direction of traffic flow.
+     *
+     * @return The traffic direction as a string.
+     */
+    public String getDirection() {
+
+        return direction;
+    }
+
+    /**
+     * Sets the direction of traffic flow.
+     *
+     * @param direction The traffic direction to be set (e.g., inbound or outbound).
+     */
+    public void setDirection(String direction) {
+
+        this.direction = direction;
+    }
+
+    /**
+     * Extracts the payload from the message context.
+     *
+     * @param messageContext the Synapse MessageContext
+     * @param config         LLM provider configuration
+     * @return extracted payload
+     */
+    private String extractPayloadFromContext(MessageContext messageContext, LLMProviderConfiguration config)
+            throws XMLStreamException, IOException {
+
+        org.apache.axis2.context.MessageContext axis2MessageContext =
+                ((Axis2MessageContext) messageContext).getAxis2MessageContext();
+
+        for (LLMProviderMetadata metadata : config.getMetadata()) {
+            if (APIConstants.AIAPIConstants.INPUT_SOURCE_PAYLOAD.equals(metadata.getInputSource())) {
+                return getPayload(axis2MessageContext);
+            }
+        }
+        return null;
+    }
+
+    /**
+     * Extracts the payload from the Axis2 message context.
+     *
+     * @param axis2MessageContext the Axis2 message context
+     * @return the extracted payload
+     */
+    private String getPayload(org.apache.axis2.context.MessageContext axis2MessageContext)
+            throws IOException, XMLStreamException {
+
+        RelayUtils.buildMessage(axis2MessageContext);
+        String contentType = (String) axis2MessageContext.getProperty(APIMgtGatewayConstants.REST_CONTENT_TYPE);
+
+        if (contentType == null) {
+            return null;
+        }
+
+        String normalizedContentType = contentType.toLowerCase();
+        if (normalizedContentType.contains(MediaType.APPLICATION_XML) || normalizedContentType.
+                contains(MediaType.TEXT_XML)) {
+            return axis2MessageContext.getEnvelope().getBody().getFirstElement().toString();
+        } else if (normalizedContentType.contains(MediaType.APPLICATION_JSON)) {
+            if (JsonUtil.hasAJsonPayload(axis2MessageContext)) {
+                return JsonUtil.jsonPayloadToString(axis2MessageContext);
+            }
+            return null;
+        } else if (normalizedContentType.contains(MediaType.TEXT_PLAIN)) {
+            return axis2MessageContext.getEnvelope().getBody().getFirstElement().getText();
+        }
+        return null;
+    }
+
+    /**
+     * Extracts query parameters from the request.
+     *
+     * @param messageContext the Synapse MessageContext
+     * @return map of query parameters
+     */
+    private Map<String, String> extractQueryParamsFromContext(MessageContext messageContext) {
+
+        org.apache.axis2.context.MessageContext axis2MessageContext =
+                ((Axis2MessageContext) messageContext).getAxis2MessageContext();
+
+        String requestPath = (String) axis2MessageContext.getProperty(RESTConstants.REST_SUB_REQUEST_PATH);
+        if (requestPath == null || requestPath.isEmpty()) {
+            log.debug("No request path available in the message context.");
+            return new HashMap<>();
+        }
+
+        return extractQueryParams(requestPath);
+    }
+
+    /**
+     * Parses the query parameters from the request path.
+     *
+     * @param requestPath the request path
+     * @return map of query parameters
+     */
+    private Map<String, String> extractQueryParams(String requestPath) {
+
+        Map<String, String> queryParams = new HashMap<>();
+        if (requestPath.contains("?")) {
+            String[] pairs = requestPath.split("\\?")[1].split("&");
+            for (String pair : pairs) {
+                String[] keyValue = pair.split("=", 2);
+                if (keyValue.length == 2) {
+                    queryParams.put(keyValue[0], keyValue[1]);
+                } else if (keyValue.length == 1) {
+                    queryParams.put(keyValue[0], "");
+                }
+            }
+        }
+        return queryParams;
+    }
+
+    /**
+     * Extracts headers from the message context.
+     *
+     * @param messageContext the Synapse MessageContext
+     * @return map of headers
+     */
+    private Map<String, String> extractHeadersFromContext(MessageContext messageContext) {
+
+        org.apache.axis2.context.MessageContext axis2MessageContext =
+                ((Axis2MessageContext) messageContext).getAxis2MessageContext();
+        return (Map<String, String>) axis2MessageContext
+                .getProperty(org.apache.axis2.context.MessageContext.TRANSPORT_HEADERS);
     }
 }
